Refactor: Update price and discount types from int to double
This commit updates the data types for price, priceAfterDiscount, and discount from `int` to `double` across multiple files.

Key changes include:
- Modified `ProductCard` widget in `custom_product_card.dart` to accept `double` for price-related parameters.
- Updated price calculation in `best_seller_screen.dart` to use `double` for `priceAfterDiscount`.
- Changed `price`, `priceAfterDiscount`, and `discount` fields to `double?` in `BestSellerDto` in `best_seller_dto.dart`.
- Changed `price`, `priceAfterDiscount`, and `discount` fields to `double?` in `BestSellerEntity` in `best_seller_entity.dart`.

+++ lib/Feature/bestSellerFeature/api/models/best_seller_dto.dart
@@ -17,11 +17,11 @@ class BestSellerDto {
   @JsonKey(name: "images")
   final List<String>? images;
   @JsonKey(name: "price")
-  final int? price;
+  final double? price;
   @JsonKey(name: "priceAfterDiscount")
-  final int? priceAfterDiscount;
+  final double? priceAfterDiscount;
   @JsonKey(name: "discount")
-  final int? discount;
+  final double? discount;
   @JsonKey(name: "rateAvg")
   final int? rateAvg;
   @JsonKey(name: "rateCount")
